+++ crates/rome_cli/src/lib.rs
@@ -80,6 +80,10 @@ impl<'app> CliSession<'app> {
                 cli_options,
                 configuration: rome_configuration,
                 paths,
+                stdin_file_path,
+                linter_enabled,
+                organize_imports_enabled,
+                formatter_enabled,
             } => commands::check::check(
                 self,
                 CheckCommandPayload {
@@ -88,6 +92,10 @@ impl<'app> CliSession<'app> {
                     cli_options,
                     configuration: rome_configuration,
                     paths,
+                    stdin_file_path,
+                    linter_enabled,
+                    organize_imports_enabled,
+                    formatter_enabled,
                 },
             ),
             RomeCommand::Ci {